+++ imports/ui/singleGenePage/GeneLink.jsx
@@ -19,19 +19,28 @@ const isNotFound = ({ gene }) => typeof gene === 'undefined';
 const NotFound = ({ transcriptId }) => <span style={{fontSize: 10}}>{transcriptId}</span>
 
 
-const GeneLink = ({ geneId, transcriptId, gene }) => {
+const GeneLink = ({ transcriptId, gene }) => {
   return (
-    <a href={`/gene/${geneId}`} style={{fontSize: 10}}>
+    <a href={`/gene/${gene.ID}`} style={{fontSize: 10}}>
       {transcriptId} {gene.attributes.Name}
     </a>
   )
 }
 
+const geneLinkDataTracker = ({ transcriptId }) => {
+  const geneSub = Meteor.subscribe('singleGene', { transcriptId });
+  const loading = !geneSub.ready();
+  const gene = Genes.findOne({ 'subfeatures.ID': transcriptId });
+  return { loading, gene, transcriptId }
+}
+
 const withConditionalRendering = compose(
+  withTracker(geneLinkDataTracker),
   withEither(isLoading, Loading),
   withEither(isNotFound, NotFound)
 )
 
+/*
 export default withTracker(({ geneId, transcriptId }) => {
   console.log(geneId, transcriptId)
   const geneSub = Meteor.subscribe('singleGene', geneId);
@@ -45,4 +54,6 @@ export default withTracker(({ geneId, transcriptId }) => {
     gene
   }
 })(withConditionalRendering(GeneLink));
+*/
 
+export default withConditionalRendering(GeneLink);