+++ imports/ui/singleGenePage/Orthogroup.jsx
@@ -43,26 +43,29 @@ const withConditionalRendering = compose(
   withEither(hasNoOrthogroup, NoOrthogroup)
 )
 
-const TreeBranches = ({nodes}) => {
+const TreeBranches = ({ nodes, chronogram = true }) => {
+  const offset = chronogram ? 0 : 50;
+  const multiplier = chronogram ? 1 : 2;
+  const value = chronogram ? 'y' : 'value';
   return nodes.map(node => {
     const d = `
-      M${node.parent.y},${node.parent.x} 
-      L${node.parent.y},${node.x} 
-      L${node.y},${node.x}`;
+      M${offset + (node.parent[value] * multiplier)},${node.parent.x} 
+      L${offset + (node.parent[value] * multiplier)},${node.x} 
+      L${offset + (node[value] * multiplier)},${node.x}`;
     return <path key={d} d={d} style={{fill: 'none', stroke: 'black', strokeWidth: 1}}/>
   })
 }
 
-const TreeNodes = ({nodes}) => {
+const TreeNodes = ({ nodes }) => {
+  console.log(nodes)
   const tipNodes = nodes.filter(node => node.data.name.length > 0);
   return tipNodes.map(node => {
     const transcriptId = node.data.name
-    const geneId = transcriptId.split('.').slice(0,-1).join('.');
     return (
-      <g key={transcriptId}>
+      <g key={`${node.x}_${node.y}_${node.data.name}`}>
         <circle cy={node.x} cx={node.y} r='3' />
         <foreignObject width='300' height='10' x={node.y + 10} y={node.x - 15}>
-          <GeneLink transcriptId={transcriptId} geneId={geneId} />
+          <GeneLink transcriptId={transcriptId} />
         </foreignObject>
       </g>
     )
@@ -76,10 +79,10 @@ const Tree = ({ tree, size, geneIds }) => {
       <ContainerDimensions>
         {
           ({width, height}) =>{
-            const svgHeight = size * 20;
+            const svgHeight = size * 12;
             const treeMap = cluster()
               .size([svgHeight, width - 300])
-              .separation((node1,node2) => { return 1})
+              .separation((node1,node2) => { return 1 })
             const treeRoot = hierarchy(tree, node => node.branchset);
 
             const treeData = treeMap(treeRoot).sum(node => node.branchLength);